feat(rdd): Implement unsafe stage marker handling and backup queue restoration
RDD.scala
- Extend isCollapserRDD to include ShuffledRDD, CoGroupedRDD, SubtractedRDD
- Improve backup iterator tracking with correlation broken guard
- Add three-tier logging strategy at outermost iterator:
Case A: All transformations safe - log with outermost iterator
Case B: Correlation broken - use backup iterator with queue restoration
Case C: No safe iterator - write UNSAFE_STAGE marker instead of empty file

Trace.scala
- Add backupQueueTL thread-local to snapshot UID queue state when backup iterator is set
- Add restoreBackupQueue() method to restore queue state before logging backup iterator
- Add UnsafeStageMarkers object with sentinel values for unsafe stages and empty partitions
- Change activeWriters from single SafeWriter to List[SafeWriter] to support multiple writers per thread
- Update commitAllLogs() to iterate through all writers for the current thread

MerkleTree.scala
- Use UnsafeStageMarkers.EMPTY_PARTITION_HASH and EMPTY_PARTITION_UID for empty partitions
- Detect UNSAFE_STAGE marker and return special leaf node with unsafe stage hash

TaskResultVerificationManager.scala
- Change addNewRunningTask tid parameter from Int to Long to support large task IDs
- Add file existence checks before Merkle tree building:
  - If only replica 1 file exists → declare replica 1 correct
  - If only replica 2 file exists → declare replica 2 correct
  - If neither exists → trigger full-task recomputation
- Add unsafe stage marker detection in Merkle tree results:
  - If both replicas are UNSAFE_STAGE → trigger full-task recomputation
  - If one is unsafe and other isn't → trigger recomputation (suspicious mismatch)
- Use UnsafeStageMarkers constants instead of hardcoded values

TaskResultGetter.scala
- Add hash storage for IndirectTaskResult (large results that spill to disk)
- Extract hash size from metricPeaks array and store via addNewResultForTid

TaskSetManager.scala
- Change taskId parameter from taskId.toInt to taskId (Long) when registering task

+++ core/src/main/scala/org/apache/spark/scheduler/TaskSetManager.scala
@@ -607,7 +607,7 @@ private[spark] class TaskSetManager(
       case _: ResultTask[_, _] | _: ShuffleMapTask =>
         val stageIndex = (taskSet.stageId, index)
         logInfo(s"[VERIFICATION REGISTER] Registering taskId=${taskId} with stageIndex=${stageIndex} (array index=${index}), executor=${execId}, host=${host}")
-        TaskResultVerificationManager.addNewRunningTask(taskId.toInt, stageIndex, taskSet.tasks(index), execId, host)
+        TaskResultVerificationManager.addNewRunningTask(taskId, stageIndex, taskSet.tasks(index), execId, host)
       case _ =>
         logDebug(s"[VERIFICATION REGISTER] Skipping registration for verification task ${task.getClass.getSimpleName} (taskId=${taskId})")
     }
